Divides the scores into batches for each worker (#3)
Each worker gets the unique set of comparisons for each ST as a single function call in turn, instead of distributing them randomly and with a function call for each pair being scores.

This significantly improves throughput on any number of cores as well as being able to scale well when running on >4 cores. The previous version gained little beyond 4 cores.